refactor(tests/user): split test file by feature
