+++ apps/website-rewrite/tests/Integration/Controllers/User/UserTestCase.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Integration\Controllers\User;
+
+use App\Tests\Integration\Controllers\ControllerTestCase;
+
+class UserTestCase extends ControllerTestCase
+{
+}
